Improve bot handler
Still not fully working but less far from doing so.

In particular, the bot uses a read timeout of 60s, which means that the
poll handler must return faster. It does mean that there is a nasty race
condition in the Swarming code when the handler returns a task just
before the timeout expires. This has to be extended upstream.